Move unrolling to common
