+++ compiler/src/iree/compiler/Codegen/Common/BUILD.bazel
@@ -170,6 +170,7 @@ iree_compiler_cc_library(
         "Transforms.cpp",
         "TypePropagation.cpp",
         "UnrollAnnotatedLoops.cpp",
+        "UnrollElementwiseOps.cpp",
         "UserConfig.cpp",
         "VectorLayoutAnalysis.cpp",
         "VectorTransferLowering.cpp",
