+++ compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPUVectorLowering.cpp
@@ -4,6 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
+#include "iree/compiler/Codegen/Common/Transforms.h"
 #include "iree/compiler/Codegen/Dialect/VectorExt/IR/VectorExtOps.h"
 #include "iree/compiler/Codegen/Dialect/VectorExt/Transforms/Transforms.h"
 #include "iree/compiler/Codegen/LLVMGPU/Passes.h"
@@ -15,7 +16,6 @@
 #include "mlir/Dialect/Math/Transforms/Passes.h"
 #include "mlir/Dialect/MemRef/Transforms/Transforms.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
-#include "mlir/Dialect/UB/IR/UBOps.h"
 #include "mlir/Dialect/Vector/IR/VectorOps.h"
 #include "mlir/Dialect/Vector/Transforms/LoweringPatterns.h"
 #include "mlir/Dialect/Vector/Transforms/VectorTransforms.h"
@@ -507,54 +507,6 @@ private:
   }
 };
 
-struct UnrollElementwiseOps final : RewritePattern {
-  UnrollElementwiseOps(MLIRContext *context, PatternBenefit benefit = 1)
-      : RewritePattern(MatchAnyOpTypeTag(), benefit, context) {}
-
-  LogicalResult matchAndRewrite(Operation *op,
-                                PatternRewriter &rewriter) const override {
-    if (!OpTrait::hasElementwiseMappableTraits(op) ||
-        op->getNumResults() != 1) {
-      return failure();
-    }
-
-    Location loc = op->getLoc();
-    VectorType dstVecTy = dyn_cast<VectorType>(op->getResult(0).getType());
-    if (!dstVecTy || dstVecTy.getRank() <= 1) {
-      return failure();
-    }
-    ArrayRef<int64_t> originalSize = dstVecTy.getShape();
-
-    Value result = ub::PoisonOp::create(rewriter, loc, dstVecTy);
-    auto subVecTy =
-        VectorType::get({originalSize.back()}, dstVecTy.getElementType());
-
-    SmallVector<int64_t> tileShape(dstVecTy.getRank() - 1, 1);
-    for (SmallVector<int64_t> offsets :
-         StaticTileOffsetRange(originalSize.drop_back(), tileShape)) {
-      // Extract from each operand.
-      SmallVector<Value> operands;
-      for (Value val : op->getOperands()) {
-        // Extract subvector if the operand is a vector. This is to handle
-        // things like arith.select which take a scalar conditional but are
-        // otherwise elementwise.
-        if (isa<VectorType>(val.getType())) {
-          val = vector::ExtractOp::create(rewriter, loc, val, offsets);
-        }
-        operands.push_back(val);
-      }
-
-      Operation *clonedOp = clone(rewriter, op, subVecTy, operands);
-      Value subResult = clonedOp->getResult(0);
-      result =
-          vector::InsertOp::create(rewriter, loc, subResult, result, offsets);
-    }
-
-    rewriter.replaceOp(op, result);
-    return success();
-  }
-};
-
 struct LLVMGPUVectorLoweringPass final
     : impl::LLVMGPUVectorLoweringPassBase<LLVMGPUVectorLoweringPass> {
   void getDependentDialects(DialectRegistry &registry) const override {
@@ -616,7 +568,7 @@ struct LLVMGPUVectorLoweringPass final
       vector::populateVectorMultiReductionUnrollingPatterns(
           contractLoweringPatterns,
           vector::VectorMultiReductionLowering::InnerReduction);
-      contractLoweringPatterns.add<UnrollElementwiseOps>(funcOp->getContext());
+      populateUnrollElementwiseOpsPatterns(contractLoweringPatterns);
       // Unroll transfer_gather ops to rank 1 and lower contiguous ones to
       // vector.transfer_read.
       IREE::VectorExt::populateVectorTransferGatherLoweringPatterns(
